requirements.txt
# Production dependencies for SV-PRO Backend
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pyjwt>=2.8.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
redis>=5.0.0
httpx>=0.27.0
aiohttp>=3.9.0
pyyaml>=6.0.1
prometheus-fastapi-instrumentator>=6.1.0
python-multipart>=0.0.9
opencv-python-headless==4.9.0.80
numpy<2
websockets>=12.0
pyzmq>=25.0
apscheduler>=3.10,<4
